CONTRIBUTING.md: Request factual descriptions
In the past, we've allowed rather non-descript summary descriptions for
projects (see the `nixpkgs-hammering` entry). In order to better comply
with the official comment guidelines laid out in the parent `awesome`
project and better serve those browsing the list, update the
contributing guidelines to specify that descriptions should also be
persuasive and factual, while providing an example format.

Additionally, fix the grammar/verb-tense for the alphabetical
requirement.